Make the site responsive and refined the code
