+++ build.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 npm run build
+sed -i '1s;^;/* eslint-disable */\n;' docs/main.main.js
+sed -i '1s;^;/* eslint-disable */\n;' docs/runtime.main.js
 npx eslint . --fix
 npx stylelint "**/*.{css,scss}" --fix --custom-syntax postcss-scss
 git add *
