+++ docs/index.html
@@ -14,7 +14,7 @@
       src="https://kit.fontawesome.com/c488aa6c77.js"
       crossorigin="anonymous"
     ></script>
-  </head>
+  <script defer src="runtime.main.js"></script><script defer src="main.main.js"></script></head>
 
   <body>
     <main>
@@ -23,34 +23,32 @@
           <nav class="header-nav">
             <a href="/" class="logo"><i class="fa-solid fa-utensils"></i></a>
 
-            <ul class="nav-links main-menu">
-              <li class="nav-item">
-                <a id="cat-beef" href="#" class="nav-link"
-                  >Beef <span class="meal-count"></span
-                ></a>
-              </li>
-              <li class="nav-item">
-                <a id="cat-seafood" href="#" class="nav-link"
-                  >Seafood <span class="meal-count"></span
-                ></a>
-              </li>
-              <li class="nav-item">
-                <a id="cat-dessert" href="#" class="nav-link"
-                  >Dessert <span class="meal-count"></span
-                ></a>
-              </li>
-            </ul>
-            <div class="menu-ico">
-              <span class="bar"></span>
-              <span class="bar"></span>
-              <span class="bar"></span>
-            </div>
+            <span class="nav-links">
+              <a id="cat-beef" href="#"
+                >Beef <span class="meal-count"></span
+              ></a>
+              <a id="cat-seafood" href="#"
+                >Seafood <span class="meal-count"></span
+              ></a>
+              <a id="cat-dessert" href="#"
+                >Dessert <span class="meal-count"></span
+              ></a>
+            </span>
           </nav>
         </div>
       </header>
       <section>
         <div id="card-container" class="container"></div>
       </section>
+
+      <!-- <section>
+      <div class="container">
+        <div id="item123" class="item">
+          <div class="placeholder">Placeholder</div>
+          <span class="like-count"></span><i class="fa-solid fa-thumbs-up item-like"></i>
+        </div>
+      </div>
+    </section> -->
 
       <footer>
         <br /><br />
