+++ docs/main.main.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 /*
  * ATTENTION: The "eval" devtool has been used (maybe by default in mode: "development").
  * This devtool is neither made for production nor for readable output files.
@@ -15,7 +16,7 @@
 /***/ ((module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../node_modules/css-loader/dist/runtime/noSourceMaps.js */ \"./node_modules/css-loader/dist/runtime/noSourceMaps.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../node_modules/css-loader/dist/runtime/api.js */ \"./node_modules/css-loader/dist/runtime/api.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__);\n// Imports\n\n\nvar ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default()((_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default()));\n// Module\n___CSS_LOADER_EXPORT___.push([module.id, \".header-nav {\\n  position: sticky;\\n  height: 50px;\\n  display: flex;\\n  padding: 5px;\\n  justify-content: space-between;\\n  padding-bottom: 50px;\\n}\\n\\n.logo {\\n  border-radius: 50%;\\n  width: 50px;\\n  height: 50px;\\n  border: 4px solid rgb(73, 73, 73);\\n  display: flex;\\n  justify-content: center;\\n  align-items: center;\\n  color: rgb(214, 214, 214);\\n  background: rgb(165, 164, 164);\\n  box-shadow: 2px 4px 8px gray;\\n  font-size: 40px;\\n}\\n\\n.for-logo-img {\\n  width: 80px;\\n  padding: 12px;\\n}\\n\\n.nav-links {\\n  display: flex;\\n  gap: 30px;\\n  padding: 2px;\\n}\\n.nav-links a {\\n  color: rgb(214, 130, 130);\\n  font-size: 28px;\\n  font-weight: 800;\\n  display: flex;\\n}\\n\\n.meal-count {\\n  font-size: 18px;\\n  height: max-content;\\n  padding: 4px;\\n  margin-left: 4px;\\n  background: rgb(214, 130, 130);\\n  color: white;\\n  border-radius: 8px;\\n}\\n\\n.main-menu {\\n  position: fixed;\\n  right: -110%;\\n  gap: 0 !important;\\n  top: 90px;\\n  flex-direction: column;\\n  background-color: rgba(255, 255, 255, 0.8);\\n  width: 100%;\\n  text-align: left !important;\\n  transition: 0.3s;\\n  align-items: flex-end !important;\\n  height: 100vh;\\n  z-index: 999;\\n  list-style-type: none !important;\\n}\\n.main-menu.active {\\n  right: 0;\\n  margin-top: 0;\\n}\\n.main-menu .nav-item .nav-link {\\n  padding: 10px;\\n  text-decoration: none;\\n}\\n\\n.menu-ico {\\n  display: block;\\n}\\n\\n.bar {\\n  display: block;\\n  width: 24px;\\n  height: 3px;\\n  margin: 5px auto;\\n  transition: all 0.3s ease;\\n  background-color: rgb(214, 130, 130);\\n}\\n\\n.menu-ico.active .bar:nth-child(2) {\\n  opacity: 0;\\n}\\n.menu-ico.active .bar:nth-child(1) {\\n  transform: translateY(8px) rotate(45deg);\\n  background-color: rgb(214, 130, 130);\\n}\\n.menu-ico.active .bar:nth-child(3) {\\n  transform: translateY(-8px) rotate(-45deg);\\n  background-color: rgb(214, 130, 130);\\n}\\n\\n.danger {\\n  color: red;\\n}\\n\\n.test {\\n  border: 2px solid red;\\n  padding: 24px;\\n  display: flex;\\n  justify-content: space-around;\\n}\\n\\nheader {\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n}\\n\\nmain {\\n  width: 100%;\\n  min-height: 100vh;\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: space-between;\\n  align-items: stretch;\\n}\\n\\nsection {\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: center;\\n  align-items: center;\\n  padding: 10px;\\n}\\n\\n.container {\\n  width: 95%;\\n}\\n\\n.placeholder {\\n  width: 300px;\\n  height: 250px;\\n  background-color: grey;\\n  display: flex;\\n  justify-content: center;\\n  align-items: center;\\n  margin-bottom: 16px;\\n}\\n\\nbutton {\\n  padding: 8px;\\n}\\n\\n.user-name {\\n  font-weight: bold;\\n}\\n\\n.comment-date {\\n  font-size: small;\\n  font-style: italic;\\n}\\n\\n.comment-count {\\n  padding: 2px 4px;\\n  border-radius: 4px;\\n  background-color: black;\\n  color: white;\\n}\\n\\n.comment-form {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n}\\n.comment-form input,\\n.comment-form textarea {\\n  width: 80%;\\n  min-height: 36px;\\n  padding: 8px;\\n}\\n.comment-form input::placeholder,\\n.comment-form textarea::placeholder {\\n  font-size: 18px;\\n}\\n.comment-form .submit-btn {\\n  width: 50%;\\n  min-width: 150px;\\n  background-color: #167ec3;\\n  border: none;\\n  border-radius: 8px;\\n  padding: 8px;\\n  font-size: 18px;\\n  color: white;\\n  font-weight: bold;\\n}\\n\\nfooter {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n}\\nfooter .container {\\n  padding: 16px;\\n  border: none;\\n  margin-bottom: 24px;\\n  border-radius: 8px;\\n  background-color: #09568a;\\n  color: white;\\n  text-align: center;\\n}\\nfooter .container a {\\n  color: white;\\n  font-weight: bold;\\n}\\n\\n@media screen and (min-width: 768px) {\\n  .container {\\n    width: 80%;\\n  }\\n  .card-grid {\\n    display: grid;\\n    grid-template-columns: repeat(2, 1fr) !important;\\n    padding: 20px;\\n    column-gap: 24px;\\n  }\\n  .menu-ico {\\n    display: none;\\n  }\\n  .main-menu {\\n    position: relative !important;\\n    display: flex;\\n    flex-direction: row !important;\\n    width: 100%;\\n    text-align: right !important;\\n    transition: 0.3s;\\n    align-items: center !important;\\n    z-index: 999;\\n    list-style-type: none;\\n    justify-content: flex-end;\\n    right: 0 !important;\\n    background: transparent !important;\\n    top: 0 !important;\\n    height: auto !important;\\n  }\\n}\\n@media screen and (min-width: 960px) {\\n  .card-grid {\\n    display: grid;\\n    grid-template-columns: repeat(3, 1fr) !important;\\n  }\\n}\\n.card-grid {\\n  display: grid;\\n  grid-template-columns: 1fr;\\n  padding: 20px;\\n  column-gap: 24px;\\n}\\n\\n.card {\\n  display: flex;\\n  margin-bottom: 20px;\\n  background-color: #fff;\\n  flex-direction: column;\\n  justify-content: space-between;\\n  align-items: center;\\n  width: 100%;\\n  height: auto;\\n  border: 1px solid #ccc;\\n  border-radius: 4px;\\n  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.1);\\n  padding-bottom: 16px;\\n  overflow: hidden;\\n}\\n.card .image-placeholder {\\n  width: 100%;\\n  aspect-ratio: 4/3;\\n  background-color: #eee;\\n}\\n.card .text-flex {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n  padding: 16px;\\n}\\n.card .text-flex p {\\n  margin: 0;\\n  font-size: 18px;\\n  font-weight: bold;\\n  text-align: center;\\n}\\n.card .for-heart {\\n  display: flex;\\n  align-items: center;\\n  margin-top: 8px;\\n  color: #e74c3c;\\n}\\n.card .for-heart i {\\n  font-style: normal;\\n  margin-left: 4px;\\n  font-size: 14px;\\n}\\n.card .for-btn {\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n  margin-top: 8px;\\n}\\n.card .comment-btn {\\n  width: 100%;\\n  height: 32px;\\n  background-color: #3498db;\\n  border: none;\\n  border-radius: 4px;\\n  color: #fff;\\n  font-size: 14px;\\n  cursor: pointer;\\n}\\n.card .comment-btn:hover {\\n  background-color: #2980b9;\\n}\\n\\n.like-count {\\n  margin-left: 4px;\\n}\\n\\n.pop-up {\\n  position: fixed;\\n  top: 50%;\\n  left: 50%;\\n  transform: translate(-50%, -50%);\\n  width: 50%;\\n  height: 90%;\\n  background-color: #fff;\\n  border-radius: 4px;\\n  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.1);\\n  padding: 16px;\\n  overflow-y: scroll;\\n}\\n.pop-up .image-placeholder {\\n  width: 100%;\\n  aspect-ratio: 5/3;\\n  border: none;\\n  background-color: #eee;\\n  object-fit: cover;\\n  object-position: center;\\n}\\n.pop-up h2 {\\n  margin: 0;\\n  font-size: 24px;\\n  font-weight: bold;\\n  text-align: center;\\n}\\n.pop-up ul {\\n  list-style: none;\\n  margin: 0;\\n  padding: 0;\\n}\\n.pop-up ul li {\\n  margin: 8px 0;\\n  font-size: 18px;\\n}\\n\\n.pop-up-bg {\\n  width: 100vw;\\n  height: 100vh;\\n  position: fixed;\\n  top: 0;\\n  left: 0;\\n  background-color: rgba(212, 212, 212, 0.5);\\n}\\n\\n.exit-btn {\\n  position: absolute;\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n  text-align: center;\\n  top: 16px;\\n  right: 36px;\\n  width: 36px;\\n  height: 36px;\\n  background-color: transparent;\\n  border: none;\\n  cursor: pointer;\\n  font-size: 20px;\\n  font-weight: bold;\\n  color: grey;\\n  border-radius: 50%;\\n  transition: all 500ms ease;\\n}\\n.exit-btn:hover {\\n  color: white;\\n  background-color: rgb(30, 30, 30);\\n}\\n\\n.comments-heading {\\n  width: max-content;\\n  padding: 12px;\\n  background-color: rgb(30, 30, 30);\\n  color: white;\\n}\\n\\n.comments-count {\\n  color: rgb(30, 30, 30);\\n  background-color: white;\\n  line-height: 1;\\n  padding: 2px 8px;\\n  border-radius: 25%;\\n}\\n\\n* {\\n  text-decoration: none;\\n  font-family: \\\"poppins\\\", sans-serif;\\n}\", \"\"]);\n// Exports\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);\n\n\n//# sourceURL=webpack://webpack-demo/./src/style/style.scss?./node_modules/css-loader/dist/cjs.js!./node_modules/sass-loader/dist/cjs.js");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../node_modules/css-loader/dist/runtime/noSourceMaps.js */ \"./node_modules/css-loader/dist/runtime/noSourceMaps.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../node_modules/css-loader/dist/runtime/api.js */ \"./node_modules/css-loader/dist/runtime/api.js\");\n/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__);\n// Imports\n\n\nvar ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default()((_node_modules_css_loader_dist_runtime_noSourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default()));\n// Module\n___CSS_LOADER_EXPORT___.push([module.id, \".header-nav {\\n  position: sticky;\\n  height: 50px;\\n  display: flex;\\n  padding: 5px;\\n  justify-content: space-between;\\n  padding-bottom: 50px;\\n}\\n\\n.logo {\\n  border-radius: 50%;\\n  width: 60px;\\n  height: 60px;\\n  border: 4px solid rgb(73, 73, 73);\\n  display: flex;\\n  justify-content: center;\\n  align-items: center;\\n  color: rgb(214, 214, 214);\\n  background: rgb(165, 164, 164);\\n  box-shadow: 2px 4px 8px gray;\\n  font-size: 40px;\\n}\\n\\n.for-logo-img {\\n  width: 80px;\\n  padding: 12px;\\n}\\n\\n.nav-links {\\n  display: flex;\\n  flex-direction: row;\\n  gap: 30px;\\n  padding: 2px;\\n}\\n.nav-links a {\\n  color: rgb(214, 130, 130);\\n  font-size: 28px;\\n  font-weight: 800;\\n  display: flex;\\n}\\n\\n.meal-count {\\n  font-size: 18px;\\n  height: max-content;\\n  padding: 4px;\\n  margin-left: 4px;\\n  background: rgb(214, 130, 130);\\n  color: white;\\n  border-radius: 8px;\\n}\\n\\n.danger {\\n  color: red;\\n}\\n\\n.test {\\n  border: 2px solid red;\\n  padding: 24px;\\n  display: flex;\\n  justify-content: space-around;\\n}\\n\\nheader {\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n}\\n\\nmain {\\n  width: 100%;\\n  min-height: 100vh;\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: space-between;\\n  align-items: stretch;\\n}\\n\\nsection {\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: center;\\n  align-items: center;\\n  padding: 10px;\\n}\\n\\n.container {\\n  width: 90%;\\n}\\n\\n.placeholder {\\n  width: 300px;\\n  height: 250px;\\n  background-color: grey;\\n  display: flex;\\n  justify-content: center;\\n  align-items: center;\\n  margin-bottom: 16px;\\n}\\n\\nbutton {\\n  padding: 8px;\\n}\\n\\n.user-name {\\n  font-weight: bold;\\n}\\n\\n.comment-date {\\n  font-size: small;\\n  font-style: italic;\\n}\\n\\n.comment-count {\\n  padding: 2px 4px;\\n  border-radius: 4px;\\n  background-color: black;\\n  color: white;\\n}\\n\\n.comment-form {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n}\\n.comment-form input,\\n.comment-form textarea {\\n  width: 80%;\\n  min-height: 36px;\\n  padding: 8px;\\n}\\n.comment-form input::placeholder,\\n.comment-form textarea::placeholder {\\n  font-size: 18px;\\n}\\n.comment-form .submit-btn {\\n  width: 50%;\\n  min-width: 150px;\\n  background-color: #167ec3;\\n  border: none;\\n  border-radius: 8px;\\n  padding: 8px;\\n  font-size: 18px;\\n  color: white;\\n  font-weight: bold;\\n}\\n\\nfooter {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n}\\nfooter .container {\\n  padding: 16px;\\n  border: 2px solid black;\\n  margin-bottom: 24px;\\n}\\n\\n@media screen and (min-width: 960px) {\\n  .container {\\n    width: 80%;\\n  }\\n}\\n.card-grid {\\n  display: grid;\\n  grid-template-columns: repeat(3, 1fr);\\n  padding: 20px;\\n  column-gap: 24px;\\n}\\n\\n.card {\\n  display: flex;\\n  margin-bottom: 20px;\\n  background-color: #fff;\\n  flex-direction: column;\\n  justify-content: space-between;\\n  align-items: center;\\n  width: 100%;\\n  height: auto;\\n  border: 1px solid #ccc;\\n  border-radius: 4px;\\n  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.1);\\n  padding-bottom: 16px;\\n  overflow: hidden;\\n}\\n.card .image-placeholder {\\n  width: 100%;\\n  aspect-ratio: 4/3;\\n  background-color: #eee;\\n}\\n.card .text-flex {\\n  display: flex;\\n  flex-direction: column;\\n  align-items: center;\\n  padding: 16px;\\n}\\n.card .text-flex p {\\n  margin: 0;\\n  font-size: 18px;\\n  font-weight: bold;\\n  text-align: center;\\n}\\n.card .for-heart {\\n  display: flex;\\n  align-items: center;\\n  margin-top: 8px;\\n  color: #e74c3c;\\n}\\n.card .for-heart i {\\n  font-style: normal;\\n  margin-left: 4px;\\n  font-size: 14px;\\n}\\n.card .for-btn {\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n  margin-top: 8px;\\n}\\n.card .comment-btn {\\n  width: 100%;\\n  height: 32px;\\n  background-color: #3498db;\\n  border: none;\\n  border-radius: 4px;\\n  color: #fff;\\n  font-size: 14px;\\n  cursor: pointer;\\n}\\n.card .comment-btn:hover {\\n  background-color: #2980b9;\\n}\\n\\n.like-count {\\n  margin-left: 4px;\\n}\\n\\n.pop-up {\\n  position: fixed;\\n  top: 50%;\\n  left: 50%;\\n  transform: translate(-50%, -50%);\\n  width: 50%;\\n  height: 90%;\\n  background-color: #fff;\\n  border-radius: 4px;\\n  box-shadow: 2px 2px 8px rgba(0, 0, 0, 0.1);\\n  padding: 16px;\\n  overflow-y: scroll;\\n}\\n.pop-up .image-placeholder {\\n  width: 100%;\\n  aspect-ratio: 5/3;\\n  border: none;\\n  background-color: #eee;\\n  object-fit: cover;\\n  object-position: center;\\n}\\n.pop-up h2 {\\n  margin: 0;\\n  font-size: 24px;\\n  font-weight: bold;\\n  text-align: center;\\n}\\n.pop-up ul {\\n  list-style: none;\\n  margin: 0;\\n  padding: 0;\\n}\\n.pop-up ul li {\\n  margin: 8px 0;\\n  font-size: 18px;\\n}\\n\\n.pop-up-bg {\\n  width: 100vw;\\n  height: 100vh;\\n  position: fixed;\\n  top: 0;\\n  left: 0;\\n  background-color: rgba(212, 212, 212, 0.5);\\n}\\n\\n.exit-btn {\\n  position: absolute;\\n  display: flex;\\n  align-items: center;\\n  justify-content: center;\\n  text-align: center;\\n  top: 16px;\\n  right: 36px;\\n  width: 36px;\\n  height: 36px;\\n  background-color: transparent;\\n  border: none;\\n  cursor: pointer;\\n  font-size: 20px;\\n  font-weight: bold;\\n  color: grey;\\n  border-radius: 50%;\\n  transition: all 500ms ease;\\n}\\n.exit-btn:hover {\\n  color: white;\\n  background-color: rgb(30, 30, 30);\\n}\\n\\n.comments-heading {\\n  width: max-content;\\n  padding: 12px;\\n  background-color: rgb(30, 30, 30);\\n  color: white;\\n}\\n\\n.comments-count {\\n  color: rgb(30, 30, 30);\\n  background-color: white;\\n  line-height: 1;\\n  padding: 2px 8px;\\n  border-radius: 25%;\\n}\\n\\n* {\\n  text-decoration: none;\\n  font-family: \\\"poppins\\\", sans-serif;\\n}\", \"\"]);\n// Exports\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);\n\n\n//# sourceURL=webpack://webpack-demo/./src/style/style.scss?./node_modules/css-loader/dist/cjs.js!./node_modules/sass-loader/dist/cjs.js");
 
 /***/ }),
 
@@ -135,7 +136,7 @@ eval("\n\n/* istanbul ignore next  */\nfunction styleTagTransform(css, styleElem
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var lodash__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! lodash */ \"./node_modules/lodash/lodash.js\");\n/* harmony import */ var lodash__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(lodash__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _style_style_scss__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./style/style.scss */ \"./src/style/style.scss\");\n/* harmony import */ var _modules_meals_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./modules/meals.js */ \"./src/modules/meals.js\");\n/* harmony import */ var _modules_nav_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./modules/nav.js */ \"./src/modules/nav.js\");\n/* harmony import */ var _modules_globals_js__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./modules/globals.js */ \"./src/modules/globals.js\");\n\n\n\n\n\nwindow.onload = () => {\n  (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.showMeals)(\"Beef\");\n\n  (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.countMeals)(\"Beef\", _modules_globals_js__WEBPACK_IMPORTED_MODULE_4__.catBeef);\n\n  _modules_globals_js__WEBPACK_IMPORTED_MODULE_4__.catSeafood.childNodes[1].style.visibility = \"hidden\";\n\n  _modules_globals_js__WEBPACK_IMPORTED_MODULE_4__.catDessert.childNodes[1].style.visibility = \"hidden\";\n\n  (0,_modules_nav_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"])();\n};\n\n\n//# sourceURL=webpack://webpack-demo/./src/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var lodash__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! lodash */ \"./node_modules/lodash/lodash.js\");\n/* harmony import */ var lodash__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(lodash__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _style_style_scss__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./style/style.scss */ \"./src/style/style.scss\");\n/* harmony import */ var _modules_meals_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./modules/meals.js */ \"./src/modules/meals.js\");\n\n\n\n// import { itemLike, likeCount } from \"./modules/like.js\";\n// import { baseLink } from \"./modules/globals.js\";\n// const items = document.querySelectorAll(\".card\");\nconst catBeef = document.getElementById('cat-beef');\nconst catSeafood = document.getElementById('cat-seafood');\nconst catDessert = document.getElementById('cat-dessert');\nconst cards = document.getElementById('show-case');\n// const mealCountSpans = document.querySelectorAll(\".meal-count\");\n// const emptySpan = () => {\n//   mealCountSpans.forEach((span) => {\n//     if (span.childNodes.length === 0) {\n//       span.style.display = \"none\";\n//     }\n//   });\n// };\nwindow.onload = () => {\n  (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.showMeals)('Beef');\n\n  (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.countMeals)('Beef', catBeef);\n\n  catSeafood.childNodes[1].style.visibility = 'hidden';\n\n  catDessert.childNodes[1].style.visibility = 'hidden';\n  catBeef.addEventListener('click', () => {\n    cards.replaceChildren();\n    (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.showMeals)('Beef');\n    (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.countMeals)('Beef', catBeef);\n    catBeef.childNodes[1].style.visibility = 'visible';\n    catSeafood.childNodes[1].innerText = '';\n    catSeafood.childNodes[1].style.visibility = 'hidden';\n    catDessert.childNodes[1].innerText = '';\n    catDessert.childNodes[1].style.visibility = 'hidden';\n  });\n  catSeafood.addEventListener('click', () => {\n    cards.replaceChildren();\n    (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.showMeals)('Seafood');\n    (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.countMeals)('Seafood', catSeafood);\n    catSeafood.childNodes[1].style.visibility = 'visible';\n    catBeef.childNodes[1].innerText = '';\n    catBeef.childNodes[1].style.visibility = 'hidden';\n    catDessert.childNodes[1].innerText = '';\n    catDessert.childNodes[1].style.visibility = 'hidden';\n  });\n  catDessert.addEventListener('click', () => {\n    cards.replaceChildren();\n    (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.showMeals)('Dessert');\n    (0,_modules_meals_js__WEBPACK_IMPORTED_MODULE_2__.countMeals)('Dessert', catDessert);\n    catDessert.childNodes[1].style.visibility = 'visible';\n    catBeef.childNodes[1].innerText = '';\n    catBeef.childNodes[1].style.visibility = 'hidden';\n    catSeafood.childNodes[1].innerText = '';\n    catSeafood.childNodes[1].style.visibility = 'hidden';\n  });\n};\n\n\n//# sourceURL=webpack://webpack-demo/./src/index.js?");
 
 /***/ }),
 
@@ -146,7 +147,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var loda
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _comments_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./comments.js */ \"./src/modules/comments.js\");\n/* harmony import */ var _globals_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./globals.js */ \"./src/modules/globals.js\");\n/* harmony import */ var _like_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./like.js */ \"./src/modules/like.js\");\n\n\n\nconst cardGrid = document.createElement(\"div\");\ncardGrid.classList.add(\"card-grid\");\ncardGrid.id = \"show-case\";\n\nconst renderCards = (arr) => {\n  arr.forEach((cardData) => {\n    const card = document.createElement(\"div\");\n    card.classList.add(\"card\");\n    card.id = cardData.idMeal;\n    const imagePlaceholder = document.createElement(\"img\");\n    imagePlaceholder.setAttribute(\"class\", \"image-placeholder\");\n    imagePlaceholder.setAttribute(\"alt\", \"\");\n    imagePlaceholder.setAttribute(\"src\", `${cardData.strMealThumb}`);\n    const textFlex = document.createElement(\"span\");\n    textFlex.setAttribute(\"class\", \"text-flex\");\n\n    const p = document.createElement(\"p\");\n    p.textContent = cardData.strMeal;\n\n    const forHeart = document.createElement(\"span\");\n    forHeart.setAttribute(\"class\", \"for-heart\");\n    const likeIco = document.createElement(\"i\");\n    likeIco.classList.add(\"fa-solid\", \"fa-heart\");\n    likeIco.id = `like-${cardData.idMeal}`;\n    likeIco.style.cursor = \"pointer\";\n    forHeart.appendChild(likeIco);\n    likeIco.addEventListener(\"click\", () => {\n      (0,_like_js__WEBPACK_IMPORTED_MODULE_2__.itemLike)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n      setTimeout(() => {\n        (0,_like_js__WEBPACK_IMPORTED_MODULE_2__.likeCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink);\n      }, 1000);\n    });\n    const likeNum = document.createElement(\"span\");\n    likeNum.classList.add(\"like-count\");\n    likeNum.id = `like-count${cardData.idMeal}`;\n    forHeart.appendChild(likeNum);\n\n    textFlex.appendChild(p);\n    textFlex.appendChild(forHeart);\n\n    const forBtn = document.createElement(\"span\");\n    forBtn.setAttribute(\"class\", \"for-btn\");\n\n    const commentBtn = document.createElement(\"button\");\n    commentBtn.setAttribute(\"class\", \"comment-btn\");\n    commentBtn.textContent = \"Comments\";\n\n    forBtn.appendChild(commentBtn);\n\n    card.appendChild(imagePlaceholder);\n    card.appendChild(textFlex);\n    card.appendChild(forBtn);\n\n    cardGrid.appendChild(card);\n\n    commentBtn.addEventListener(\"click\", () => {\n      document.body.classList.add(\"pop-up-open\");\n      const popUpBg = document.createElement(\"div\");\n      popUpBg.classList.add(\"pop-up-bg\");\n      const popUp = document.createElement(\"div\");\n      popUp.classList.add(\"pop-up\");\n      popUp.id = `pop-${cardData.idMeal}`;\n      popUp.innerHTML = `\n        <img class=\"image-placeholder\" src='${cardData.strMealThumb}' alt='${cardData.strMeal}'/>\n        <h2>${cardData.strMeal}</h2>\n        <h3 class=\"comments-heading\">Comments <span class=\"comments-count\" id=\"#comment-count-${cardData.idMeal}\"></span></h3>\n        <div class=\"show-comments\"></ul>\n        `;\n\n      const exitBtn = document.createElement(\"button\");\n      exitBtn.classList.add(\"exit-btn\");\n      exitBtn.innerHTML = '<i class=\"fa-solid fa-xmark\"></i>';\n      popUpBg.appendChild(exitBtn);\n      const commentForm = document.createElement(\"div\");\n      commentForm.classList.add(\"form-cintainer\");\n      commentForm.innerHTML = `<form name=\"add-comment\" class=\"comment-form\" id=\"comment-${cardData.idMeal}\">\n      <input\n        id=\"username-${cardData.idMeal}\"\n        type=\"text\"\n        placeholder=\"Name (required)\"\n        name=\"name\"\n        class=\"name\"\n        required\n      />\n      <br />\n      <textarea\n        id=\"user-comment-${cardData.idMeal}\"\n        placeholder=\"Your comment (required)\"\n        name=\"comment\"\n        class=\"comment\"\n        required\n      ></textarea\n      ><br />\n      <button id=\"submit-btn-${cardData.idMeal}\" class=\"submit-btn\" type=\"button\">Submit</button>\n    </form>`;\n      popUp.appendChild(commentForm);\n      popUpBg.appendChild(popUp);\n      document.body.appendChild(popUpBg);\n\n      exitBtn.addEventListener(\"click\", () => {\n        document.body.removeChild(popUpBg);\n        document.body.classList.remove(\"pop-up-open\");\n      });\n      (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.showComment)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n      const submitBtn = document.getElementById(\n        `submit-btn-${cardData.idMeal}`\n      );\n      submitBtn.addEventListener(\"click\", (e) => {\n        e.preventDefault();\n        (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.addComment)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n        //   commentForm.reset();\n        setTimeout(() => {\n          (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.showComment)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n          (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.showCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n        }, 1000);\n      });\n      (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.showCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n    });\n  });\n};\n\nconst cardContainer = document.getElementById(\"card-container\");\ncardContainer.appendChild(cardGrid);\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (renderCards);\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/card.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _comments_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./comments.js */ \"./src/modules/comments.js\");\n/* harmony import */ var _globals_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./globals.js */ \"./src/modules/globals.js\");\n/* harmony import */ var _like_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./like.js */ \"./src/modules/like.js\");\n\n\n\n// const cardGrid = document.querySelector('.middle');\nconst cardGrid = document.createElement('div');\ncardGrid.classList.add('card-grid');\ncardGrid.id = 'show-case';\n// cardsSection.appendChild(cardGrid);\nconst renderCards = (arr) => {\n  arr.forEach((cardData) => {\n    const card = document.createElement('div');\n    card.classList.add('card');\n    card.id = cardData.idMeal;\n    const imagePlaceholder = document.createElement('img');\n    imagePlaceholder.setAttribute('class', 'image-placeholder');\n    imagePlaceholder.setAttribute('alt', '');\n    imagePlaceholder.setAttribute('src', `${cardData.strMealThumb}`);\n    const textFlex = document.createElement('span');\n    textFlex.setAttribute('class', 'text-flex');\n\n    const p = document.createElement('p');\n    p.textContent = cardData.strMeal;\n\n    const forHeart = document.createElement('span');\n    forHeart.setAttribute('class', 'for-heart');\n    const likeIco = document.createElement('i');\n    likeIco.classList.add('fa-solid', 'fa-heart');\n    likeIco.id = `like-${cardData.idMeal}`;\n    likeIco.style.cursor = 'pointer';\n    forHeart.appendChild(likeIco);\n    likeIco.addEventListener('click', () => {\n      (0,_like_js__WEBPACK_IMPORTED_MODULE_2__.itemLike)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n      setTimeout(() => {\n        (0,_like_js__WEBPACK_IMPORTED_MODULE_2__.likeCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink);\n      }, 1000);\n    });\n    const likeNum = document.createElement('span');\n    likeNum.classList.add('like-count');\n    likeNum.id = `like-count${cardData.idMeal}`;\n    forHeart.appendChild(likeNum);\n\n    textFlex.appendChild(p);\n    textFlex.appendChild(forHeart);\n\n    const forBtn = document.createElement('span');\n    forBtn.setAttribute('class', 'for-btn');\n\n    const commentBtn = document.createElement('button');\n    commentBtn.setAttribute('class', 'comment-btn');\n    commentBtn.textContent = 'Comments';\n\n    // commentBtn.addEventListener('click', () => {\n    //   const comments = cardData.comments.map((comment) => `<li>${comment}</li>`).join('');\n    //   alert(`<h2>${cardData.name}</h2><ul>${comments}</ul>`);\n    // });\n\n    forBtn.appendChild(commentBtn);\n\n    card.appendChild(imagePlaceholder);\n    card.appendChild(textFlex);\n    card.appendChild(forBtn);\n\n    cardGrid.appendChild(card);\n\n    commentBtn.addEventListener('click', () => {\n      document.body.classList.add('pop-up-open');\n      const popUpBg = document.createElement('div');\n      popUpBg.classList.add('pop-up-bg');\n      const popUp = document.createElement('div');\n      popUp.classList.add('pop-up');\n      popUp.id = `pop-${cardData.idMeal}`;\n      popUp.innerHTML = `\n        <img class=\"image-placeholder\" src='${cardData.strMealThumb}' alt='${cardData.strMeal}'/>\n        <h2>${cardData.strMeal}</h2>\n        <h3 class=\"comments-heading\">Comments <span class=\"comments-count\" id=\"#comment-count-${cardData.idMeal}\"></span></h3>\n        <div class=\"show-comments\"></ul>\n        `;\n\n      const exitBtn = document.createElement('button');\n      exitBtn.classList.add('exit-btn');\n      exitBtn.innerHTML = '<i class=\"fa-solid fa-xmark\"></i>';\n      popUpBg.appendChild(exitBtn);\n      const commentForm = document.createElement('div');\n      commentForm.classList.add('form-cintainer');\n      commentForm.innerHTML = `<form name=\"add-comment\" class=\"comment-form\" id=\"comment-${cardData.idMeal}\">\n      <input\n        id=\"username-${cardData.idMeal}\"\n        type=\"text\"\n        placeholder=\"Name (required)\"\n        name=\"name\"\n        class=\"name\"\n        required\n      />\n      <br />\n      <textarea\n        id=\"user-comment-${cardData.idMeal}\"\n        placeholder=\"Your comment (required)\"\n        name=\"comment\"\n        class=\"comment\"\n        required\n      ></textarea\n      ><br />\n      <button id=\"submit-btn-${cardData.idMeal}\" class=\"submit-btn\" type=\"button\">Submit</button>\n    </form>`;\n      popUp.appendChild(commentForm);\n      popUpBg.appendChild(popUp);\n      document.body.appendChild(popUpBg);\n\n      exitBtn.addEventListener('click', () => {\n        document.body.removeChild(popUpBg);\n        document.body.classList.remove('pop-up-open');\n      });\n      (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.showComment)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n      const submitBtn = document.getElementById(\n        `submit-btn-${cardData.idMeal}`,\n      );\n      submitBtn.addEventListener('click', (e) => {\n        e.preventDefault();\n        (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.addComment)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n        //   commentForm.reset();\n        setTimeout(() => {\n          (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.showComment)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n          (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.showCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n        }, 1000);\n      });\n      (0,_comments_js__WEBPACK_IMPORTED_MODULE_0__.showCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink, cardData.idMeal);\n    });\n  });\n};\n\nconst cardContainer = document.getElementById('card-container');\ncardContainer.appendChild(cardGrid);\n\n// document.body.appendChild(cardGrid);\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (renderCards);\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/card.js?");
 
 /***/ }),
 
@@ -168,7 +169,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"appId\": () => (/* binding */ appId),\n/* harmony export */   \"baseLink\": () => (/* binding */ baseLink),\n/* harmony export */   \"catBeef\": () => (/* binding */ catBeef),\n/* harmony export */   \"catDessert\": () => (/* binding */ catDessert),\n/* harmony export */   \"catSeafood\": () => (/* binding */ catSeafood),\n/* harmony export */   \"menuIco\": () => (/* binding */ menuIco),\n/* harmony export */   \"navMenu\": () => (/* binding */ navMenu)\n/* harmony export */ });\nconst appId = \"No6xjeOV6L9eg8TkvJgU\";\nconst baseLink = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${appId}`;\nconst menuIco = document.querySelector(\".menu-ico\");\nconst navMenu = document.querySelector(\".main-menu\");\nconst catBeef = document.getElementById(\"cat-beef\");\nconst catSeafood = document.getElementById(\"cat-seafood\");\nconst catDessert = document.getElementById(\"cat-dessert\");\n\n\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/globals.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"appId\": () => (/* binding */ appId),\n/* harmony export */   \"baseLink\": () => (/* binding */ baseLink)\n/* harmony export */ });\nconst appId = 'No6xjeOV6L9eg8TkvJgU';\nconst baseLink = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${appId}`;\n\n\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/globals.js?");
 
 /***/ }),
 
@@ -191,17 +192,6 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"countMeals\": () => (/* binding */ countMeals),\n/* harmony export */   \"showMeals\": () => (/* binding */ showMeals)\n/* harmony export */ });\n/* harmony import */ var _card_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./card.js */ \"./src/modules/card.js\");\n/* harmony import */ var _globals_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./globals.js */ \"./src/modules/globals.js\");\n/* harmony import */ var _like_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./like.js */ \"./src/modules/like.js\");\n\n\n\n\nconst showMeals = async (category) => {\n  const response = await fetch(\n    `https://www.themealdb.com/api/json/v1/1/filter.php?c=${category}`,\n  );\n  await response.json().then((food) => {\n    const foods = food.meals;\n    (0,_card_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"])(foods);\n  });\n  (0,_like_js__WEBPACK_IMPORTED_MODULE_2__.likeCount)(_globals_js__WEBPACK_IMPORTED_MODULE_1__.baseLink);\n};\nconst mealsCount = (arr) => {\n  let result = 0;\n  if (arr.length === undefined) {\n    result = 0;\n  } else {\n    result = arr.length;\n  }\n  return result;\n};\nconst countMeals = async (category, menuItem) => {\n  const response = await fetch(\n    `https://www.themealdb.com/api/json/v1/1/filter.php?c=${category}`,\n  );\n  await response.json().then((food) => {\n    const foods = food.meals;\n    menuItem.childNodes[1].innerText = mealsCount(foods);\n  });\n};\n\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/meals.js?");
-
-/***/ }),
-
-/***/ "./src/modules/nav.js":
-/*!****************************!*\
-  !*** ./src/modules/nav.js ***!
-  \****************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _meals_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./meals.js */ \"./src/modules/meals.js\");\n/* harmony import */ var _globals_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./globals.js */ \"./src/modules/globals.js\");\n\n\n\nconst navigation = () => {\n  const cards = document.getElementById(\"show-case\");\n  _globals_js__WEBPACK_IMPORTED_MODULE_1__.catBeef.addEventListener(\"click\", () => {\n    cards.replaceChildren();\n    (0,_meals_js__WEBPACK_IMPORTED_MODULE_0__.showMeals)(\"Beef\");\n    (0,_meals_js__WEBPACK_IMPORTED_MODULE_0__.countMeals)(\"Beef\", _globals_js__WEBPACK_IMPORTED_MODULE_1__.catBeef);\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catBeef.childNodes[1].style.visibility = \"visible\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catSeafood.childNodes[1].innerText = \"\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catSeafood.childNodes[1].style.visibility = \"hidden\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catDessert.childNodes[1].innerText = \"\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catDessert.childNodes[1].style.visibility = \"hidden\";\n  });\n  _globals_js__WEBPACK_IMPORTED_MODULE_1__.catSeafood.addEventListener(\"click\", () => {\n    cards.replaceChildren();\n    (0,_meals_js__WEBPACK_IMPORTED_MODULE_0__.showMeals)(\"Seafood\");\n    (0,_meals_js__WEBPACK_IMPORTED_MODULE_0__.countMeals)(\"Seafood\", _globals_js__WEBPACK_IMPORTED_MODULE_1__.catSeafood);\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catSeafood.childNodes[1].style.visibility = \"visible\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catBeef.childNodes[1].innerText = \"\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catBeef.childNodes[1].style.visibility = \"hidden\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catDessert.childNodes[1].innerText = \"\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catDessert.childNodes[1].style.visibility = \"hidden\";\n  });\n  _globals_js__WEBPACK_IMPORTED_MODULE_1__.catDessert.addEventListener(\"click\", () => {\n    cards.replaceChildren();\n    (0,_meals_js__WEBPACK_IMPORTED_MODULE_0__.showMeals)(\"Dessert\");\n    (0,_meals_js__WEBPACK_IMPORTED_MODULE_0__.countMeals)(\"Dessert\", _globals_js__WEBPACK_IMPORTED_MODULE_1__.catDessert);\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catDessert.childNodes[1].style.visibility = \"visible\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catBeef.childNodes[1].innerText = \"\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catBeef.childNodes[1].style.visibility = \"hidden\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catSeafood.childNodes[1].innerText = \"\";\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.catSeafood.childNodes[1].style.visibility = \"hidden\";\n  });\n\n  _globals_js__WEBPACK_IMPORTED_MODULE_1__.menuIco.addEventListener(\"click\", () => {\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.menuIco.classList.toggle(\"active\");\n    _globals_js__WEBPACK_IMPORTED_MODULE_1__.navMenu.classList.toggle(\"active\");\n  });\n\n  document.querySelectorAll(\".nav-link\").forEach((n) =>\n    n.addEventListener(\"click\", () => {\n      _globals_js__WEBPACK_IMPORTED_MODULE_1__.menuIco.classList.remove(\"active\");\n      _globals_js__WEBPACK_IMPORTED_MODULE_1__.navMenu.classList.remove(\"active\");\n    })\n  );\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (navigation);\n\n\n//# sourceURL=webpack://webpack-demo/./src/modules/nav.js?");
 
 /***/ })
