+++ src/modules/globals.js
@@ -1,4 +1,11 @@
 const appId = 'No6xjeOV6L9eg8TkvJgU';
 const baseLink = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${appId}`;
+const menuIco = document.querySelector('.menu-ico');
+const navMenu = document.querySelector('.main-menu');
+const catBeef = document.getElementById('cat-beef');
+const catSeafood = document.getElementById('cat-seafood');
+const catDessert = document.getElementById('cat-dessert');
 
-export { appId, baseLink };
+export {
+  appId, baseLink, menuIco, navMenu, catBeef, catSeafood, catDessert,
+};
