+++ src/style/_common.scss
@@ -35,7 +35,7 @@ section {
 }
 
 .container {
-  width: 90%;
+  width: 95%;
 }
 
 .placeholder {
