+++ src/style/_nav.scss
@@ -8,10 +8,9 @@
 }
 
 .logo {
-  // margin-left: 8%;
   border-radius: 50%;
-  width: 60px;
-  height: 60px;
+  width: 50px;
+  height: 50px;
   border: 4px solid rgb(73, 73, 73);
   display: flex;
   justify-content: center;
@@ -29,7 +28,6 @@
 
 .nav-links {
   display: flex;
-  flex-direction: row;
   gap: 30px;
   // margin-right: 10%;
   padding: 2px;
@@ -50,4 +48,63 @@
   background: rgb(214, 130, 130);
   color: white;
   border-radius: 8px;
+}
+
+.main-menu {
+  position: fixed;
+  right: -110%;
+  gap: 0 !important;
+  top: 90px;
+  flex-direction: column;
+  background-color: rgba($color: #fff, $alpha: 0.8);
+  width: 100%;
+  text-align: left !important;
+  transition: 0.3s;
+  align-items: flex-end !important;
+  height: 100vh;
+  z-index: 999;
+  list-style-type: none !important;
+
+  &.active {
+    right: 0;
+    margin-top: 0;
+  }
+
+  .nav-item {
+    .nav-link {
+      padding: 10px;
+      text-decoration: none;
+    }
+  }
+}
+
+.menu-ico {
+  display: block;
+}
+
+.bar {
+  display: block;
+  width: 24px;
+  height: 3px;
+  margin: 5px auto;
+  transition: all 0.3s ease;
+  background-color: rgb(214, 130, 130);
+}
+
+.menu-ico.active {
+  .bar {
+    &:nth-child(2) {
+      opacity: 0;
+    }
+
+    &:nth-child(1) {
+      transform: translateY(8px) rotate(45deg);
+      background-color: rgb(214, 130, 130);
+    }
+
+    &:nth-child(3) {
+      transform: translateY(-8px) rotate(-45deg);
+      background-color: rgb(214, 130, 130);
+    }
+  }
 }
